LayoutTests/platform/ios-sim-deprecated/http/tests/security/cross-frame-access-name-getter-expected.txt

CONSOLE MESSAGE: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
 
----- tests for getting properties by name -----

PASS: canGet('targetWindow.testiframe') should be 'true' and is.
PASS: canGet('targetWindow.testimage') should be 'false' and is.
